+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+**Release date:** `N/A`
+
+### Changed
+
+- Separate `PickleIterator.readBytes` to `PickleIterator.readBytes` & `PickleIterator.readAs`.
+
 ## 1.0.4
 
 **Release date:** `2023-06-23`